src/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure.ai.documentintelligence
azure-ai-textanalytics
azure.core
azure-functions
azure-functions-durable
langchain
langdetect
python-dotenv
semantic_kernel
azure.storage.blob
tiktoken